requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
llama_index==0.10.20
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.31.0
gunicorn==21.0.1
Flask-Session==0.8.0
python-telegram-bot==21.3
bs4==0.0.2
supervisor==4.2.2